[build] extract mbox source into its own module
