+++ build.gradle
@@ -42,6 +42,7 @@ subprojects {
         scalaBinaryVersion = '2.12'
         githubApiVersion = '1.109'
         okHttpVersion = '3.13.1'
+        mime4jVersion = '0.8.3'
         log4jVersion = '2.13.1'
     }
