+++ import/build.gradle
@@ -16,12 +16,15 @@ dependencies {
     flinkShadowJar "org.kohsuke:github-api:${githubApiVersion}"
     flinkShadowJar "com.squareup.okhttp3:okhttp-urlconnection:${okHttpVersion}"
 
+    implementation project(path: ':source-mbox')
+    flinkShadowJar project(path: ':source-mbox', transitive: false)
+    flinkShadowJar "org.apache.james:apache-mime4j-core:${mime4jVersion}"
+    flinkShadowJar "org.apache.james:apache-mime4j-dom:${mime4jVersion}"
+    flinkShadowJar "org.apache.james:apache-mime4j-mbox-iterator:${mime4jVersion}"
+
     flinkShadowJar "org.apache.flink:flink-connector-elasticsearch7_${scalaBinaryVersion}:${flinkVersion}"
     flinkShadowJar "org.apache.flink:flink-connector-kafka_${scalaBinaryVersion}:${flinkVersion}"
     flinkShadowJar "org.apache.flink:flink-json:${flinkVersion}"
-    flinkShadowJar "org.apache.james:apache-mime4j-core:0.8.3"
-    flinkShadowJar "org.apache.james:apache-mime4j-dom:0.8.3"
-    flinkShadowJar "org.apache.james:apache-mime4j-mbox-iterator:0.8.3"
 }
 
 publishing {
