+++ import/src/main/java/com/ververica/platform/FlinkPullRequestsToKafka.java
@@ -1,7 +1,7 @@
 package com.ververica.platform;
 
 import static com.ververica.platform.FlinkMailingListToKafka.DATE_OR_DATETIME_FORMATTER;
-import static com.ververica.platform.io.source.GithubSource.EVALUATION_ZONE;
+import static com.ververica.platform.Utils.EVALUATION_ZONE;
 
 import com.ververica.platform.entities.PullRequest;
 import com.ververica.platform.io.source.GithubPullRequestSource;
