+++ source-mbox/build.gradle
@@ -1,0 +1,16 @@
+plugins {
+    id 'java-library'
+}
+
+dependencies {
+    // --------------------------------------------------------------
+    // Compile-time dependencies that should NOT be part of the
+    // shadow jar and are provided in the lib folder of Flink
+    // --------------------------------------------------------------
+    api "org.apache.flink:flink-java:${flinkVersion}"
+    api "org.apache.flink:flink-streaming-java_${scalaBinaryVersion}:${flinkVersion}"
+
+    flinkShadowJar "org.apache.james:apache-mime4j-core:${mime4jVersion}"
+    flinkShadowJar "org.apache.james:apache-mime4j-dom:${mime4jVersion}"
+    flinkShadowJar "org.apache.james:apache-mime4j-mbox-iterator:${mime4jVersion}"
+}
